requirements.txt
aniso8601==7.0.0
coverage==4.5.4
Django==2.2.5
django-graphql-jwt==0.2.2
django-nose==1.4.6
factory-boy==2.12.0
Faker==2.0.1
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
nose==1.3.7
promise==2.2.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
